+++ out/extension.js
@@ -33,8 +33,15 @@ const fs = __importStar(require("fs"));
 let config = vscode.workspace.getConfiguration('stardew-pets');
 let webview;
 let extensionStorageFolder = '';
+//Enemy enums
+const EnemySpecies = {
+    slime: ['iron', 'tiger'],
+    bug: ['normal', 'normal dangerous', 'armored', 'armored dangerous'],
+    crab: ['rock', 'rock dangerous', 'lava', 'lava dangerous', 'iridium', 'truffle', 'stickbug', 'magma cap'],
+    golem: ['stone', 'stone dangerous', 'iridium', 'wilderness'],
+};
 //Pet enums (species & names)
-const Species = {
+const PetSpecies = {
     cat: ['black', 'gray', 'orange', 'white', 'yellow', 'purple'],
     dog: ['blonde', 'gray', 'brown', 'dark brown', 'light brown', 'purple'],
     dino: [],
@@ -195,7 +202,7 @@ function activate(context) {
     //Add pet
     const commandAddPet = vscode.commands.registerCommand('stardew-pets.addPet', async () => {
         //Ask for a specie
-        const specie = await vscode.window.showQuickPick(Object.keys(Species), {
+        const specie = await vscode.window.showQuickPick(Object.keys(PetSpecies), {
             title: 'Select a pet',
             placeHolder: 'pet',
         });
@@ -203,8 +210,8 @@ function activate(context) {
             return;
         //Ask for a variant
         let variants = Array();
-        for (let i = 0; i < Species[specie].length; i++) {
-            const variant = Species[specie][i];
+        for (let i = 0; i < PetSpecies[specie].length; i++) {
+            const variant = PetSpecies[specie][i];
             //Get adult/baby start index
             let index = variant.indexOf(' adult');
             if (index == -1)
@@ -399,6 +406,7 @@ class WebViewProvider {
                 <script src="${this.getUri(webview, 'util.js')}"></script>
                 <script src="${this.getUri(webview, 'base.js')}"></script>
                 <script src="${this.getUri(webview, 'pets.js')}"></script>
+                <script src="${this.getUri(webview, 'enemies.js')}"></script>
                 <script src="${this.getUri(webview, 'main.js')}"></script>
             </body>
             </html>